// packages/cel/src/lib/protogen/cel/expr/eval_pb.ts
export const file_cel_expr_eval: GenFile = /*@__PURE__*/
  fileDesc("ChNjZWwvZXhwci9ldmFsLnByb3RvEghjZWwuZXhwciKEAQoJRXZhbFN0YXRlEiMKBnZhbHVlcxgBIAMoCzITLmNlbC5leHByLkV4cHJWYWx1ZRIrCgdyZXN1bHRzGAMgAygLMhouY2VsLmV4cHIuRXZhbFN0YXRlLlJlc3VsdBolCgZSZXN1bHQSDAoEZXhwchgBIAEoAxINCgV2YWx1ZRgCIAEoAyKDAQoJRXhwclZhbHVlEiAKBXZhbHVlGAEgASgLMg8uY2VsLmV4cHIuVmFsdWVIABIjCgVlcnJvchgCIAEoCzISLmNlbC5leHByLkVycm9yU2V0SAASJwoHdW5rbm93bhgDIAEoCzIULmNlbC5leHByLlVua25vd25TZXRIAEIGCgRraW5kIiwKCEVycm9yU2V0EiAKBmVycm9ycxgBIAMoCzIQLmNlbC5leHByLlN0YXR1cyJOCgZTdGF0dXMSDAoEY29kZRgBIAEoBRIPCgdtZXNzYWdlGAIgASgJEiUKB2RldGFpbHMYAyADKAsyFC5nb29nbGUucHJvdG9idWYuQW55IhsKClVua25vd25TZXQSDQoFZXhwcnMYASADKANCLAoMZGV2LmNlbC5leHByQglFdmFsUHJvdG9QAVoMY2VsLmRldi9leHBy+AEBYgZwcm90bzM", [file_cel_expr_value, file_google_protobuf_any]);
// ...
export type ExprValue = Message<"cel.expr.ExprValue"> & {
  /**
   * An expression can resolve to a value, error or unknown.
   *
   * @generated from oneof cel.expr.ExprValue.kind
   */
  kind: {
    /**
     * @generated from field: cel.expr.Value value = 1;
     */
    value: Value;
    case: "value";
  } | {
    /**
     * The set of errors in the critical path of evaluation.
     *
     * Only errors in the critical path are included. For example,
     * `(<error1> || true) && <error2>` will only result in `<error2>`,
     * while `<error1> || <error2>` will result in both `<error1>` and
     * `<error2>`.
     *
     * Errors cause by the presence of other errors are not included in the
     * set. For example `<error1>.foo`, `foo(<error1>)`, and `<error1> + 1` will
     * only result in `<error1>`.
     *
     * Multiple errors *might* be included when evaluation could result
     * in different errors. For example `<error1> + <error2>` and
     * `foo(<error1>, <error2>)` may result in `<error1>`, `<error2>` or both.
     * The exact subset of errors included for this case is unspecified and
     * depends on the implementation details of the evaluator.
     *
     * @generated from field: cel.expr.ErrorSet error = 2;
     */
    value: ErrorSet;
    case: "error";
  } | {
    /**
     * The set of unknowns in the critical path of evaluation.
     *
     * Unknown behaves identically to Error with regards to propagation.
     * Specifically, only unknowns in the critical path are included, unknowns
     * caused by the presence of other unknowns are not included, and multiple
     * unknowns *might* be included when evaluation could result in
     * different unknowns. For example:
     *
     *     (<unknown[1]> || true) && <unknown[2]> -> <unknown[2]>
     *     <unknown[1]> || <unknown[2]> -> <unknown[1,2]>
     *     <unknown[1]>.foo -> <unknown[1]>
     *     foo(<unknown[1]>) -> <unknown[1]>
     *     <unknown[1]> + <unknown[2]> -> <unknown[1]> or <unknown[2[>
     *
     * Unknown takes precedence over Error in cases where a `Value` can short
     * circuit the result:
     *
     *     <error> || <unknown> -> <unknown>
     *     <error> && <unknown> -> <unknown>
     *
     * Errors take precedence in all other cases:
     *
     *     <unknown> + <error> -> <error>
     *     foo(<unknown>, <error>) -> <error>
     *
     * @generated from field: cel.expr.UnknownSet unknown = 3;
     */
    value: UnknownSet;
    case: "unknown";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message cel.expr.ExprValue.
 * Use `create(ExprValueSchema)` to create a new message.
 */
export const ExprValueSchema: GenMessage<ExprValue> = /*@__PURE__*/
  messageDesc(file_cel_expr_eval, 1);

/**
 * A set of errors.
 *
 * The errors included depend on the context. See `ExprValue.error`.
 *
 * @generated from message cel.expr.ErrorSet
 */
export type ErrorSet = Message<"cel.expr.ErrorSet"> & {
  /**
   * Errors that could come up during evaluation phase.
   *
   * @generated from field: repeated cel.expr.Status errors = 1;
   */
  errors: Status[];
};
// ...
export type Status = Message<"cel.expr.Status"> & {
  /**
   * The status code, which should be an enum value of [google.rpc.Code][].
   *
   * @generated from field: int32 code = 1;
   */
  code: number;

  /**
   * A developer-facing error message, which should be in English. Any
   * user-facing error message should be localized and sent in the
   * [Status.details][] field, or localized by the client.
   *
   * @generated from field: string message = 2;
   */
  message: string;

  /**
   * A list of messages that carry the error details.  There is a common set of
   * message types for APIs to use.
   *
   * @generated from field: repeated google.protobuf.Any details = 3;
   */
  details: Any[];
};

/**
 * Describes the message cel.expr.Status.
 * Use `create(StatusSchema)` to create a new message.
 */
export const StatusSchema: GenMessage<Status> = /*@__PURE__*/
  messageDesc(file_cel_expr_eval, 3);

/**
 * A set of expressions for which the value is unknown.
 *
 * The unknowns included depend on the context. See `ExprValue.unknown`.
 *
 * @generated from message cel.expr.UnknownSet
 */
export type UnknownSet = Message<"cel.expr.UnknownSet"> & {
  /**
   * The ids of the expressions with unknown values.
   *
   * @generated from field: repeated int64 exprs = 1;
   */
  exprs: bigint[];
};